Se arreglaron bugs con contacto y grupo

+++ controllers/contacto.js
@@ -1,6 +1,9 @@
 const ContactoServices = require("../services/contacto.js");
+const UsuarioServices = require("../services/usuario.js");
 
 const service = new ContactoServices();
+
+const usuarioService = new UsuarioServices();
 
 /**
  * @author Bernardo de la Sierra
@@ -31,10 +34,37 @@ class ContactoController {
   }
 
   /**
+   * @author Julio Meza
+   * @version 1.0.1
+   * @license Gp
+   * @params Sin parametros
+   * @description Funcion que te da todos los contactos de un usuario
+   */
+  async getAllContactosUsuario(req, res) {
+    try {
+
+      const id = req.params.id;
+
+      if (!Number.isInteger(parseInt(id))) {
+        return res.status(500).json({ message: "El Id necesita ser entero" });
+      }
+
+      const contactos = await service.getContactosDeUsuario(id);
+      return res.status(200).json({ data: contactos });
+    } catch (err) {
+      return res
+        .status(500)
+        .json({ message: `Error al obtener los contactos. Err: ${err}` });
+    }
+  }
+
+
+
+  /**
    * @author Bernardo de la Sierra
    * @version 1.0.1
    * @license Gp
-   * @params {int} - id Identificador unico del contacto
+   * @params Sin parametros
    * @description Funcion que te da determinado contacto por id
    */
   async getContacto(req, res) {
@@ -53,7 +83,7 @@ class ContactoController {
     } catch (err) {
       return res
         .status(500)
-        .json({ message: `Error al obtener los contacto. Err: ${err}` });
+        .json({ message: `Error al obtener el contacto. Err: ${err}` });
     }
   }
 
@@ -70,6 +100,12 @@ class ContactoController {
       return res.status(500).json({ message: "El Id necesita ser entero" });
     }
     try {
+
+      const contacto = await service.getContacto(id);
+      if (!contacto) {
+        return res.status(404).json({ message: "No se encontró el contacto" });
+      }
+
       await service.deleteContacto(id);
       return res
         .status(200)
@@ -78,6 +114,57 @@ class ContactoController {
       return res
         .status(500)
         .json({ message: `Error al obtener los contactos. Err: ${err}` });
+    }
+  }
+
+
+   /**
+   * @author Bernardo de la Sierra
+   * @version 1.0.1
+   * @license Gp
+   * @params {int} - idUsuarioActual Identificador del usuario que esta registrando el contacto
+   * @params {string} - correo es el correo del usuario a añadir en la relacion
+   * @description  Funcion que crea las relaciones de contactos
+   */
+   async addContacto(req, res) {
+    const { idUsuarioActual, correo } = req.body;
+
+    try {
+
+      //buscar si el usuario que se desea añadir como contacto existe
+      const usuarioAgregado = await usuarioService.getUsuarioPorCorreo(correo);
+
+      //si no existe el usuario lanzar un error
+      if (!usuarioAgregado) {
+        return res
+          .status(400)
+          .json({ message: "El usuario no se encuentra registrado" });
+      }
+
+      //si el usuario que se agrega como contacto es el mismo usuario lanzar un error
+      if(usuarioAgregado.idUsuario == idUsuarioActual){
+        return res
+        .stauts(400)
+        .json({ message: "No se puede agregar el mismo usuario"});
+      }
+
+       //buscar que no se haya registrado ya el contacto
+       const contacto = await service.getContactoPorUsuarios(idUsuarioActual, usuarioAgregado.idUsuario)
+       if(contacto) {
+          return res
+          .status(400)
+          .json({message: "El contacto ya se ha registrado"});
+       }
+
+      //crear el contacto
+      const contactoCreado = await service.addContacto(
+        idUsuarioActual,
+        usuarioAgregado.idUsuario
+      );
+
+      res.status(200).json(contactoCreado);
+    } catch (error) {
+      res.status(500).json({ error: error.message });
     }
   }
 }
