+++ controllers/grupo.js
@@ -54,7 +54,7 @@ class GrupoController {
     } catch (err) {
       return res
         .status(500)
-        .json({ message: `Error al obtener los grupo. Err: ${err}` });
+        .json({ message: `Error al obtener los grupos. Err: ${err}` });
     }
   }
 
@@ -89,6 +89,12 @@ class GrupoController {
       return res.status(500).json({ message: "El Id necesita ser entero" });
     }
     try {
+
+      const grupo = await service.getGrupo(id);
+      if (!grupo) {
+        return res.status(404).json({ message: "No se encontró el grupo" });
+      }
+
       await service.deleteGrupo(id);
       return res
         .status(200)
@@ -108,24 +114,24 @@ class GrupoController {
    * @params {int} - idGrupo Unico del grupo
    * @description Funcion para darle registro a determinado grupo
    */
-  async addUsuarioGrupo(req, res) {
-    const { idUsuarioAgregado, idGrupo } = req.body;
+  async addUsuarioToGrupo(req, res) {
+    const { idUsuario, idGrupo } = req.body;
 
     try {
-      //buscar si el correo está registrado para evitar registrar de nuevo
-      const id = await usuarioService.getUsuario(idUsuarioAgregado);
+      //buscar si el usuario está registrado
+      const id = await usuarioService.getUsuario(idUsuario);
       if (!id) {
         return res
           .status(400)
           .json({ message: "El usuario no se encuentra registrado" });
       }
-      console.log(id);
-      const grupo = await service.createUsuarioGrupo(id, idGrupo);
+
+      const grupo = await service.addUsuarioGrupo(idUsuario, idGrupo);
       return res.status(200).json({ grupo });
     } catch (err) {
       return res
         .status(500)
-        .json({ message: `Error al crear el grupo. Err: ${err}` });
+        .json({ message: `Error al añadir el usuario al grupo. Err: ${err}` });
     }
   }
 
@@ -136,15 +142,14 @@ class GrupoController {
    * @params {int} - id Identificador unico del grupo
    * @description Funcion que nos va a permitir ver todos los contactos en un grupo
    */
-  async getUsuarioPorGrupo(req, res) {
+  async getUsuariosGrupo(req, res) {
     const id = req.params.id;
     // Verificamos que el id no sea un string
     if (!Number.isInteger(parseInt(id))) {
-      return res.status(500).json({ message: "El Id necesita ser entero" });
+      return res.status(500).json({ message: "El id necesita ser entero" });
     }
     try {
-      console.log(id);
-      const contactoGrupo = await service.getUsuarioPorGrupo(id);
+      const contactoGrupo = await service.getUsuariosGrupo(id);
       if (contactoGrupo) {
         return res.status(200).json({ data: contactoGrupo });
       } else {
